log_generators/pyproject.toml
[tool.poetry]
name = "log-generators"
version = "0.1.0"
description = ""
authors = ["Mihajlo Madic <mihajlo.madic@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
tomli = "^2.1.0"
validators = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
